backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
aioboto3==9.6.0
    # via -r requirements.in
aiobotocore[boto3]==2.3.0
    # via
    #   -r requirements.in
    #   aioboto3
aiofiles==0.8.0
    # via -r requirements.in
aiohttp==3.8.1
    # via -r requirements.in
aioitertools==0.10.0
    # via -r requirements.in
aioredis==2.0.1
    # via -r requirements.in
aiosignal==1.2.0
    # via
    #   -r requirements.in
    #   aiohttp
anyio==3.6.1
    # via
    #   -r requirements.in
    #   httpcore
    #   starlette
    #   watchgod
asgiref==3.5.2
    # via
    #   -r requirements.in
    #   uvicorn
async-timeout==4.0.2
    # via
    #   -r requirements.in
    #   aiohttp
    #   aioredis
    #   redis
asyncpg==0.25.0
    # via -r requirements.in
attrs==22.1.0
    # via
    #   -r requirements.in
    #   aiohttp
bcrypt==3.2.2
    # via -r requirements.in
black==22.3.0
    # via -r requirements.in
boto3==1.21.21
    # via -r requirements.in
botocore==1.24.21
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
certifi==2022.6.15
    # via
    #   -r requirements.in
    #   httpcore
    #   requests
cffi==1.15.0
    # via
    #   -r requirements.in
    #   bcrypt
    #   cryptography
charset-normalizer==2.0.12
    # via
    #   -r requirements.in
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
    #   uvicorn
cryptography==37.0.3
    # via -r requirements.in
deprecated==1.2.13
    # via
    #   -r requirements.in
    #   redis
dnspython==2.2.1
    # via -r requirements.in
ecdsa==0.17.0
    # via
    #   -r requirements.in
    #   python-jose
fastapi==0.78.0
    # via -r requirements.in
frozenlist==1.3.1
    # via
    #   -r requirements.in
    #   aiohttp
    #   aiosignal
gunicorn==20.1.0
    # via -r requirements.in
h11==0.12.0
    # via
    #   -r requirements.in
    #   httpcore
    #   uvicorn
httpcore==0.15.0
    # via -r requirements.in
httptools==0.4.0
    # via -r requirements.in
idna==3.3
    # via
    #   -r requirements.in
    #   anyio
    #   requests
    #   yarl
itsdangerous==2.1.2
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   -r requirements.in
    #   boto3
    #   botocore
markupsafe==2.1.1
    # via -r requirements.in
multidict==6.0.2
    # via
    #   -r requirements.in
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via
    #   -r requirements.in
    #   black
orjson==3.7.2
    # via -r requirements.in
packaging==21.3
    # via
    #   -r requirements.in
    #   redis
passlib==1.7.4
    # via -r requirements.in
pathspec==0.9.0
    # via
    #   -r requirements.in
    #   black
pep517==0.12.0
    # via
    #   -r requirements.in
    #   pip-tools
pillow==9.2.0
    # via -r requirements.in
pip-tools==6.6.2
    # via -r requirements.in
platformdirs==2.5.2
    # via
    #   -r requirements.in
    #   black
pyasn1==0.4.8
    # via
    #   -r requirements.in
    #   python-jose
    #   rsa
pycparser==2.21
    # via
    #   -r requirements.in
    #   cffi
pydantic==1.9.1
    # via
    #   -r requirements.in
    #   fastapi
pyparsing==3.0.9
    # via
    #   -r requirements.in
    #   packaging
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   botocore
python-dotenv==0.20.0
    # via -r requirements.in
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
redis==4.3.3
    # via -r requirements.in
requests==2.28.0
    # via -r requirements.in
rfc3986==1.5.0
    # via -r requirements.in
rsa==4.8
    # via
    #   -r requirements.in
    #   python-jose
s3transfer==0.5.2
    # via
    #   -r requirements.in
    #   boto3
six==1.16.0
    # via
    #   -r requirements.in
    #   ecdsa
    #   python-dateutil
    #   python-multipart
sniffio==1.2.0
    # via
    #   -r requirements.in
    #   anyio
    #   httpcore
sortedcontainers==2.4.0
    # via -r requirements.in
starlette==0.19.1
    # via
    #   -r requirements.in
    #   fastapi
tomli==2.0.1
    # via
    #   -r requirements.in
    #   black
    #   pep517
typing-extensions==4.2.0
    # via
    #   -r requirements.in
    #   aioredis
    #   pydantic
ujson==5.3.0
    # via -r requirements.in
urllib3==1.26.9
    # via
    #   -r requirements.in
    #   botocore
    #   requests
uvicorn==0.17.6
    # via -r requirements.in
uvloop==0.16.0
    # via -r requirements.in
watchgod==0.8.2
    # via -r requirements.in
websockets==10.3
    # via -r requirements.in
wheel==0.37.1
    # via pip-tools
wrapt==1.14.1
    # via
    #   -r requirements.in
    #   deprecated
yarl==1.8.1
    # via
    #   -r requirements.in
    #   aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
